opex: moved the logic from the stale cases report into a helpers file, then made a new stale cases file that calls it
